Dynamic width for combo boxes on android
Make sure that combo boxes are wide enough to fit all of the items in the
model. This is primarily a fix for android since the org.kde.desktop
theme already inherits this behavior from the underlying widgets
implementation, but it actually applies to all of the built-in Qt themes.

Fixed #22