Reset article list when clicking title
This mimics the behavior of clicking the feed name in the feed list.

Unfortunately this does not animate the transition: animating a list
scroll can be an *extremely* expensive operation, potentially
instantiating the delegate for *every item in the list*.

If this turns out to be a usability problem we can probably fake a scroll
animation with shaders, although combining it with the appropriate item
displacement animations could be tricky.

fixes #27