[Android] Support icon coloring
Icon coloring on desktop is supported by KIconThemes, which colors icons
by overriding the QIcon loading process to insert CSS code into SVG icons
on the fly. KIconThemes is not buildable on android and has too many
KDE dependencies, so I've copy-pasted the core svg theming code into
a new AndroidStylePluginIconLoader class. This is not a fully generic
solution, but it works for our purposes.

fixes #18