Don't add feeds from invalid OPML file
importOpml was adding feeds immediately as the QXMLStreamReader was
reading the file. This minimized memory usage but resulted in partial
imports if an error was encountered while reading the file.

Instead, store a list of the feeds in the OPML file and save them all
after the entire file has been successfully read.