Group all database writes into a single transaction per event
Automatically group write operations into transactions. This is
achieved by pushing a low-priority event onto the event queue to
commit the transaction.  Since the commit event has a lower priority
than the async-write events, we will have at most one transaction per
run of the event loop, and everything will be commited to the database
on idle (or on application exit).