Refactor icon provider to support overlapping requests
The image provider for feed icons no longer makes multiple network
requests for feed icons when an icon that is currently being loaded
from the network is requested again.

This avoids timing problems when requesting an icon immediately after
icon discovery, or when scrolling the feed list up and down rapidly.

Fixes #33