Avoid internal int overflow in QNetworkDiskCache
We want to cache all the favicons, regardless of size, since we know that
we are going to need them on every application start.  Previously we were
using INT64_MAX as the cache limit, but Qt uses multiples of the cache
size internally, so we need a smaller value. 1GB should be enough.