Implement a custom feed discovery algorithm
We've been using the Syndication library's built-in regex-based feed
discovery, which is quite unreliable.  This one uses the gumbo parser,
which we were already using to generate the article content model.

As part of this change, the UpdatableFeed class is reimplemented to
hook into the Syndication::ParserCollection directly instead of using
Loader/DataRetriever, since the Loader API doesn't provide access
to the raw data.  Some glue code for interfacing with the gumbo parser
is also moved into the core lib.