Don't try to open article pages while property updates are in progress
AbstractArticleListPage::openChild gets called when the current item
in the list view changes, but that includes during e.g. the construction
and destruction of the page. That often happens in the middle of some
other operation on the page stack, which leads to erratic behavior.

To avoid this, we push the page manipulation code into the event queue
instead of calling it directly.