Hack to disable GlobalDrawer scroll view
GlobalDrawer wraps everything in a ScrollView, but we're filling it
with a ListView that is also scrollable.  Since nested scroll views cause
event handling glitches, and we can't access the scroll view directly,
we use a hack to disable the enclosing flickable.

This fixes #35 for the time being. In the long term, we should move
away from GlobalDrawer to our own drawer layout, since GlobalDrawer is
not really flexible enough to do what we want to do with it without a
lot of workarounds (see also issue #15).

fixes #35