Workaround for infinite loop in about page layout
AboutPage is freezing with certain font settings when trying to wrap
the homepage URL (#46). AboutPage is a stock component so I can't directly
modify the layout parameters, but padding the end of the URL string with
whitespace keeps it from trying to wrap in the middle of the URL and
triggering this bug.