Build Article Content Model Asynchronously
Rather than laying out all of the article content during the QML loading
process, we parse the content and post a series of low-priority events
that do the layout for each block separately.

This improves responsiveness when loading large articles, although blocks
are currently only broken at text/image boundaries, so articles that
present a long wall of text with no images might still be sluggish.

This has the possibility to cause regressions, so it shouldn't be picked
to the release series until after 1.0.

fixes #49