Use the old Kirigami.ScrollablePage
Kirigami replaced the old ScrollablePage implementation with one based
on QQC2.ScrollView.  The new implementation has various glitches with
the refresh indicator and the scroll bar, so I copied the old version
into the project for now.

Right now I'm only adding this to the release branch so that we're not
shipping a broken product.