Switch from RichText to StyledText in article list
This brings back support for eliding text and limiting line count on the
headline, at the expense of support for certain HTML content in titles
(and a stricter parser, but that's a non-issue here since our titles get
extracted from XML).

If the limited HTML subset becomes a problem we can add some preprocessing
to the titles when extracting them from the feed.  However, HTML
formatting in article titles is rare, and I don't see a problem with
ignoring some of it. The primary use case here is HTML entities, and
StyledText supports the most common ones.

Fixes #77