Introduce ToggleItem for article list
Rework the article list based on a new ToggleItem class, which switches
between two different list items based on the read value. This currently
provides a smooth transition when the read state and unread state have
different heights, and also lays the groundwork for adding a swipe
gesture between the two states.