Run QReadable on background thread
the QReadable script can take a significant amount of time to run, so we
don't want to block the UI while running it.

Note that we create our own worker thread instead of using e.g. QRunnable
so that we can create and re-use a single instance of QReadable, which
is not thread-safe.