Avoid conflicting animations when replacing the root page
We were previously clearing the page row and then pushing the root page,
but in cases where clearing the page row produces a transition (e.g. when
there are more pages than fit on the screen), this results in overlapping
animations and a broken end state.

This change replaces the clear-then-push sequence with a new pushRoot
function which calls push if the row is empty and replace if it is not.

fixes #85