Experimental support for Tiny Tiny RSS remote backend
using the following environment variables:

- SYNDIC_STORAGE=ttrss
- SYNDIC_TTRSS_ENDPOINT (the api endpoint for the ttrss instance)
- SYNDIC_TTRSS_USERNAME
- SYNDIC_TTRSS_PASSWORD

This implementation essentially turns the application into a thin UI layer
for  the remote server: storage operations are mapped directly to RPC
calls; updating feeds is a no-op, and all of the feed loading and parsing
is done by the remote server.

This is currently *very* rough and nowhere near production ready.