Stop abusing signals for window-global actions
There were a couple of places where we had created signals on page
components for the sole purpose of triggering specific function calls
on the application window. I'm not sure why this seemed like a good
idea at the time, but it adds needless complexity to the control flow.

Replace these signals with direct calls to member functions of the
application window.