Propagate timestamp when scheduling stale feed
Feeds now use the timestamp passed to UpdateScheduler::schedule as their
update time if they were already stale when inserted. This ensures that
all initially-stale feeds stay on the same schedule. Before this fix,
feeds could end up offset by a second if the process of initially
scheduling them happened to cross a tick of the clock.

This issue was previously causing intermittent test failures in
testStaleFeedGetsUpdatedWhenScheduled. That test has been modified so
that it now always fails if the timestamp is not passed through. It
previously only failed when the test happened to run at an inconvenient
clock time.