Show image placeholders an animate image loading
This avoids the situation where an article loads, you begin reading it,
and then a slow-loading image triggers an abrupt relayout. We now show
a placeholder where images are meant to be.  If the loading takes
longer than a threshold time (currently hardcoded at 250ms) we show
a transition to the final image when it loads.