Work around Cloudflare sometimes blocking image requests
Cloudflare sometimes produces a challenge page for image requests over
HTTP/1.1. This is particularly an issue for sites that leave Under
Attack mode on all the time.

Try to detect when this might be happening and retry with HTTP/2 enabled.
(NB: we don't turn on HTTP/2 by default because Qt5's HTTP/2
implementation is unreasonably slow)