Avoid creating multiple instances of the "All Items" feed
The all items feed is a fairly heavy object which recieves and re-emits
signals from all of the other feeds. Avoid unnecessarily recreating it
by having a single shared instance provided by the context.

With this change the AllItemsFeed class becomes an internal implementation
detail of the new Context::allItemsFeed() function.