Make sure article list is sorted
The article list previously displayed articles in whatever order they
were retrieved from the database, which was fine because our database
queries are written to return articles in the correct order. But the
category feeds make multiple queries to get all of their articles so
the model needs to sort them.