Try to guess image sizes for loading placeholder
When an image tag doesn't include an explicit size, we need to guess the
size.  Previously we just used the size of the loading icon, but we can
often do better: the URL of an image often includes the size as part of
the file name.

With this change we now try to guess the image size based on the URL
before falling back to the default size. Our size-extraction algorithm
eliminates image-load relayouts for several popular feeds.