Don't reload article page for every article (#135)
Until now the article page was designed to display a single article, and
the article list created a new one every time an article
was selected. Swiping between articles was implemented with a hack where
swiping to an empty item would trigger navigation on the list.

This replaces that entire mechanism with a horizontal list of articles
that uses the same model as the list page. which has several advantages:
* Efficiency: no more unnecessarily creating and destroying pages
* Touch interaction: we can show the actual content of the article being
swiped to
* Preloading: We can use ListView::cacheBuffer to pre-render the next
page while the current one is being read
* Fixes bug #56, which was almost impossible to avoid with the old
design, and probably avoids some other similar glitches where Kirigami's
async page loading gets batched with our async text rendering