Use declarative properties to populate pagerow
Introduce a new PageController class which consolidates all of the logic
for pushing and popping pages from the page row into a single plage. The
PageController has a property naming the root page component, and each
page has a PageControl.nextPage attached property defining the page to
appear after it in the page row.

This fixes #147 and makes it easier to consistently implement column
view related bugfixes in the future.