Allow building without QReadable for testing purposes
Introduce the WITHOUT_QREADABLE build option to build feedcore without
a dependency on QReadable. When built in this manner, attempts to fetch
readable content will immediately fail.

This is very much not recommended for builds that will be used by humans,
since it renders the readability-related features of the application
completely non-functional, but it is sometimes convenient for development
and testing purposes.