More prominently display the source feed for articles
We previously appended a link to the source feed at the end of the byline,
but that could often be hidden if the byline was long or the window was
narrow.

This redesigns the article view a bit so that the source feed is displayed
with its icon above the title.

Fixes #155