Automatically fetch more recommended items as the existing ones are read
This ensures that a user who only interacts with the application through
the highlights page will eventually see all unread articles. This is
important if the highlights page is going to be the default start page
going forward, otherwise naive users may end up never seeing some of the
articles in their collection.