Move web content pre-fetch from updater to automation rule
Moves the existing logic for pre-fetching web page text with readability
out of the updater into the automation engine. The update can now finish
while web content is still being downloaded in the background.

Since the web content prefetch is simply populating a cache, there is no
reason to make the user wait for it to complete before browsing; the worst
case scenario is just a cache miss.

This change makes it a bit more urgent that we create a user interface
for managing the automation engine and its tasks (issue #189).

Fixes #253