Fall back to scraping links from the web page if a feed can't be found
Adds a new flag for a feed that is actually a web page. These are
processed using our existing gumbo-based HTML processing infrastructure
to identify likely article links based on common URL patterns. The
result is a basic links-only feed.

The web page flag is not set by the user, rather it gets set
automatically when feed discovery is attempted but fails to yield a
suitable feed.